django/pyproject.toml
[tool.poetry]
name = "arva-django"
version = "0.1.0"
description = ""
authors = ["Charlie Koch <charles.w.koch@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
Django = "*"
django-concurrency = "*"
psycopg = { version = "*", extras = ["binary"] }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
